requirements.txt
Flask==2.2.3
python-pptx==0.6.21
Pillow==9.4.0
Werkzeug==2.2.3
flask-cors==3.0.10
markdown==3.4.4